Restructurizing CS-files
