+++ Voxel Tycoon Open Library/Reflection/VTOLReflectionHelpers.cs
@@ -2,7 +2,7 @@
 using System;
 using System.Reflection;
 
-namespace VTOL
+namespace VTOL.Reflection
 {
     /// <summary>
     /// Class with methods to help with reflection
