+++ tests/VTOL.Tests/VTOLReflectionHelpersTests.cs
@@ -1,5 +1,6 @@
 ﻿using NUnit.Framework;
 using System;
+using VTOL.Reflection;
 
 namespace VTOL
 {
